Add `get_model_state` to get validated doc.
This adds `get_model_state` which returns the entire doc state as the
pydantic model pass in as `Model=` during instantiation.

It's effectively the other side of `apply_update`.

If the doc has no Model defined it will raise a RuntimeError.
If the doc is invalid it will raise a pydantic ValidationError.

Also, update `apply_update` to use `model_validate` instead of
Model(**value). This is more of a "style" thing.
See https://github.com/pydantic/pydantic/discussions/9676

Also add ruff to the test dependencies and format.